requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-redis==6.0.0
factory_boy==3.3.3
Faker==37.5.3
flower==2.0.1
humanize==4.12.3
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
